Moved 6 loglevel overloads to T4 template
Moved 6 loglevel overloads to T4 template. T4 is default included in
VS2012+
